fix tests attempt

+++ src/haswell/avx2_convert_utf16_to_latin1.cpp
@@ -60,7 +60,7 @@ std::pair<result, char*> avx2_convert_utf16_to_latin1_with_errors(const char16_t
                 if(word <= 0xff) {
                     *latin1_output++ = char(word);
                 } else {
-                    return std::make_pair(result{error_code::TOO_LARGE, buf - start + k}, latin1_output);
+                    return std::make_pair(result{error_code::TOO_LARGE, (size_t)(buf - start + k)}, latin1_output);
                 }
             }
             buf += 16;
